+++ frictionless/plugins/inline.py
@@ -21,11 +21,12 @@ class InlinePlugin(Plugin):
 
     def create_file(self, file):
         if not file.scheme and not file.format and file.memory:
-            types = (list, typing.Iterator, typing.Generator)
-            if callable(file.data) or isinstance(file.data, types):
-                file.scheme = ""
-                file.format = "inline"
-                return file
+            if not hasattr(file.data, "read"):
+                types = (list, typing.Iterator, typing.Generator)
+                if callable(file.data) or isinstance(file.data, types):
+                    file.scheme = ""
+                    file.format = "inline"
+                    return file
 
     def create_dialect(self, resource, *, descriptor):
         if resource.format == "inline":