+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Here described only the breaking and most significant changes. The full changelog and documentation for all released versions could be found in nicely formatted [commit history](https://github.com/frictionlessdata/frictionless-py/commits/master).
 
+## v4.0.0a15
+
+- Updated loaders (#658) (BREAKING)
+    - Renamed `filelike` loader to `stream` loader
+    - Migrated from `text` loader to `buffer` loader
+
 ## v4.0.0a14
 
 - Improve transform API (#657) (BREAKING)